Accumulate typechecking methods in impl TypeCheckContext
All methods from `impl Namespace` that also take `Engines` as a parameter
have been moved into `impl TypeCheckContext`.

List of changes:

- `ctx.namespace.find_method_for_type` -> `ctx.find_method_for_type`
- `ctx.namespace.find_constant_for_type` -> `ctx.find_constant_for_type`
- `ctx.namespace.get_items_for_type_and_trait_name` -> `ctx.get_items_for_type_and_trait_name`
- `ctx.namespace.insert_trait_implementation` -> `ctx.insert_trait_implementation`
- `ctx.namespace.variant_import` -> `ctx.variant_import`
- `ctx.namespace.item_import` -> `ctx.item_import`
- `ctx.namespace.self_import` -> `ctx.self_import`
- `ctx.namespace.variant_star_import` -> `ctx.variant_star_import`
- `ctx.namespace.star_import` -> `ctx.star_import`
- `ctx.namespace.find_items_for_type` -> `ctx.find_items_for_type`
- `ctx.namespace.resolve_call_path_with_visibility_check` -> `ctx.resolve_call_path_with_visibility_check`
- `ctx.namespace.resolve_type_with_self` -> `ctx.resolve_type_with_self` (and the previous specialized version in `ctx` removed)
- `ctx.namespace.resolve_type_without_self` -> `ctx.resolve_type_without_self` (and the previous specialized version in `ctx` removed)

Partially resolves issue #5050
