+++ sway-core/src/semantic_analysis/ast_node/declaration/constant.rs
@@ -37,6 +37,7 @@ impl ty::TyConstantDecl {
             .resolve_type_with_self(
                 handler,
                 type_ascription.type_id,
+                ctx.self_type(),
                 &type_ascription.span,
                 EnforceTypeArguments::No,
                 None,
