+++ sway-core/src/semantic_analysis/ast_node/declaration/enum.rs
@@ -75,6 +75,7 @@ impl ty::TyEnumVariant {
             .resolve_type_with_self(
                 handler,
                 type_argument.type_id,
+                ctx.self_type(),
                 &type_argument.span,
                 EnforceTypeArguments::Yes,
                 None,
