+++ sway-core/src/semantic_analysis/ast_node/declaration/function.rs
@@ -117,6 +117,7 @@ impl ty::TyFunctionDecl {
             .resolve_type_with_self(
                 handler,
                 return_type.type_id,
+                ctx.self_type(),
                 &return_type.span,
                 EnforceTypeArguments::Yes,
                 None,
