+++ sway-core/src/semantic_analysis/ast_node/declaration/struct.rs
@@ -71,6 +71,7 @@ impl ty::TyStructField {
             .resolve_type_with_self(
                 handler,
                 type_argument.type_id,
+                ctx.self_type(),
                 &type_argument.span,
                 EnforceTypeArguments::Yes,
                 None,
